Get tests runing properly
I think the test appeared to be working because one of my step methods
had the name test_.... Fixed that and nothing broke.  The colons after
when: Then: etc, seemed to be breaking the feature file so it wasn't
matching any steps. Removed those and stuff broke which was a good sign!

Subsequently added some detail to the tests so I have a placeholder
for starting a server and returing an object with a .url that we
could visit. Might change that later, but we can at least get
splinter running a browser now.