Add signal handler to close gracefully on ^c
Despit the doom and gloom here:
http://stackoverflow.com/questions/17101502/how-to-stop-the-tornado-web-server-with-ctrlc
This appears to work consistently from the comand line in a few tests.

I've switched from using IOLoop.current() to Instance for the stat
call but this ought to be okay in the context of the main app since
I'n not likely to try  and call this code from tests or anything which
will create separate ioloop instances.