Add app fixture to start threaded server
Turns out I can't use the pytest-tornado plugin's fixture because its
not threaded, its just async. Fine if you're using the async http
client. But no good for splinter browser powered by synchronous python
runing in the thread. So consulting:
http://stackoverflow.com/questions/5375220/how-do-i-stop-tornado-web-server

I built a start and stop method that run the ioloop in a thread
which seems to be okay for running tests.