Refactor app fixture
The port number was hard coded in the fixture. This needs to come from
the runner (where it is currenly hard coded). Making start() return the
url rather than the thread object makes things a bit more sane: we got
something we're actually interested in rather than something we need
just for house keeping. And remove a failure vector where we pass the
wrong thread object to stop. Let them manage it. It's a module level
global which might be winced at, but its maanged by the methods and not
exposed. A bit like a class variable.